lambdas/src/pyproject.toml
[project]
name = "gcb-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.55.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.26.4",
    "agno>=2.1.0",
    "openai>=2.0.1",
    "boto3>=1.40.60",
    "boto3-stubs[essential]>=1.40.61",
    "botocore>=1.40.60",
    "mypy-boto3-s3>=1.40.61",
]
